+++ zulassung2/src/main/java/mops/zulassung2/model/Account.java
@@ -1,0 +1,15 @@
+package mops.zulassung2.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.Set;
+
+@Getter
+@AllArgsConstructor
+public class Account {
+    private final String name;
+    private final String email;
+    private final String image;
+    private final Set<String> roles;
+}
